Numpy compatibility issues.

+++ test/test_PLY.py
@@ -10,9 +10,9 @@ def test():
     for i in np.linspace(-10, 10, 5):
         for j in np.linspace(-10, 10, 5):
             color = [
-                np.int(np.random.rand()*255),
-                np.int(np.random.rand()*255),
-                np.int(np.random.rand()*255)
+                np.int64(np.random.rand()*255),
+                np.int64(np.random.rand()*255),
+                np.int64(np.random.rand()*255)
             ]
             cad.draw_a_ray([0., 0., 0.], [i, j, 50.], color=color)
     cad.save_PLY(savefn='out.ply')
